update to 64 bits

+++ lib/evaluate.c
@@ -2,12 +2,12 @@
 #include <stdint.h>
 #include <stdlib.h>
 
-static inline int32_t max(int x, int y) {
+static inline int64_t max(int x, int y) {
     return x > y ? x : y;
 }
 
-int32_t evaluate(int32_t *sol, int32_t **mj_table, int32_t num_machines, int32_t num_jobs) {
-    int32_t *machine_cache = (int32_t*)calloc(num_machines, sizeof(int32_t));
+int64_t evaluate(int64_t *sol, int64_t **mj_table, int64_t num_machines, int64_t num_jobs) {
+    int64_t *machine_cache = (int64_t*)calloc(num_machines, sizeof(int64_t));
 
     for (int i = 0; i < num_jobs; i++) {
         for (int machine_id = 0; machine_id < num_machines; machine_id++) {
@@ -19,7 +19,7 @@ int32_t evaluate(int32_t *sol, int32_t **mj_table, int32_t num_machines, int32_t
         }
     }
 
-    int32_t makespan = machine_cache[num_machines - 1];
+    int64_t makespan = machine_cache[num_machines - 1];
     free(machine_cache);
     return makespan;
 }
