+++ utils.py
@@ -7,10 +7,10 @@ class PFSProblem:
         self.lib = ctypes.CDLL(lib_dir)
         self.lib.evaluate.restype = ctypes.c_int32
         self.lib.evaluate.argtypes = [
-            np.ctypeslib.ndpointer(dtype=np.int32, ndim=1, flags="C_CONTIGUOUS"),
-            np.ctypeslib.ndpointer(dtype=np.int32, ndim=2, flags="C_CONTIGUOUS"),
-            ctypes.c_int32,
-            ctypes.c_int32,
+            np.ctypeslib.ndpointer(dtype=np.int64, ndim=1, flags="C_CONTIGUOUS"),
+            np.ctypeslib.ndpointer(dtype=np.int64, ndim=2, flags="C_CONTIGUOUS"),
+            ctypes.c_int64,
+            ctypes.c_int64,
         ]
 
         with open(input_dir) as input_file:
